% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{apply_heatmap_by_scenario}
\alias{apply_heatmap_by_scenario}
\title{apply_map_by_scenario}
\usage{
apply_heatmap_by_scenario(dat, hot, title_txt)
}
\arguments{
\item{dat}{A \code{tibble} returned by [\code{\link[=make_heatmap_data]{make_heatmap_data()}}].}

\item{hot}{A boolean specifying whether to use a hot or cool colorscale.
Defaults to TRUE.}

\item{title_txt}{A string to use as the plot title.}
}
\value{
A \code{cowplot} object.
}
\description{
A workaround to using [\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}]. Using this function
gives plots with unique legends for easier interpretation.
}
\examples{
\dontrun{
1+1
}
}
