% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{apply_map_by_scenario}
\alias{apply_map_by_scenario}
\title{apply_map_by_scenario}
\usage{
apply_map_by_scenario(dat, shp, hot, title_txt)
}
\arguments{
\item{dat}{A \code{tibble} returned by [\code{\link[=make_map_data]{make_map_data()}}].}

\item{hot}{A boolean specifying whether to use a hot or cool colorscale.
Defaults to TRUE.}

\item{title_txt}{A string to use as the plot title.}
}
\value{
A \code{cowplot} object.
}
\description{
A workaround to using [\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}]. You can't facet_wrap
with [\code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}}], so this functio  is a workaround allowing you to
effectively facet_wrap with a \code{geom_sf}.
}
\examples{
\dontrun{
1+1
}
}
