% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{cmip_monthly_to_change}
\alias{cmip_monthly_to_change}
\title{cmip_monthly_to_change}
\usage{
cmip_monthly_to_change(files, pattern, fun, tsfm, is.annual, out_dir)
}
\arguments{
\item{files}{list of cmip6 files to preprocess.}

\item{pattern}{The pattern to use to filter \code{files}}

\item{fun}{The function to use to aggregate rasters with (usually either
"mean" or "sum").}

\item{tsfm}{A function to apply to the rasters to convert units (optional).}

\item{is.annual}{Boolean of whether the CMIP files are already aggregated to a annual timescale.}

\item{out_dir}{The directory to save the files out to.}
}
\description{
cmip_monthly_to_change
}
