% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{cmip_ppt_interannual_variability}
\alias{cmip_ppt_interannual_variability}
\title{cmip_ppt_interannual_variability}
\usage{
cmip_ppt_interannual_variability(files, out_dir)
}
\arguments{
\item{files}{List of CMIP6 files to be filtered.}

\item{out_dir}{The directory to save the results out to.}
}
\value{
Save out an \code{rds} file of interannual variability
}
\description{
Calculate interannual precipitation variability
}
\examples{
\notrun{
1+1
}
}
