% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crop_to_roi}
\alias{crop_to_roi}
\title{crop_to_roi}
\usage{
crop_to_roi(data_dir, out_dir, roi, variables)
}
\arguments{
\item{data_dir}{The directory containing raw gridmet data.}

\item{out_dir}{The directory to save the cropped results to.}

\item{roi}{A \code{sf} object outlining the region of interest to crop.}

\item{variables}{A vector of gridMET variables to crop (e.g. \code{c("pr", "tmmx")}).}
}
\value{
Nothing is returned. New files are saved to \code{out_dir}.
}
\description{
A function to crop a directory of raw gridMET data to a specified region of interest
}
\examples{
\dontrun{
crop_to_roi("~/data_dir/raw", "~/data_dir/processed", normals::mt, c("pr", "tmmx"))
}
}
