% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{divide}
\alias{divide}
\title{An \code{sf} object that splits Montana east and west of the continental divide.}
\format{
A geometry set with one feature.
}
\usage{
divide
}
\description{
An \code{sf} object that splits Montana east and west of the continental divide.
}
\keyword{datasets}
