% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_gridmet.R
\name{fetch_gridmet}
\alias{fetch_gridmet}
\title{fetch_gridmet}
\usage{
fetch_gridmet(data_dir, variables, start_year = 1991, end_year = 2020)
}
\arguments{
\item{data_dir}{Directory to save data to. If the directory doesn't exist,
it will be created. Raw data will be saved to \code{data_dir/raw}. An additional
\code{data_dir/processed} will be created to save processed data in later.}

\item{variables}{A character vector of gridMET variables to download, options
to choose from are \code{c("erc",  "etr",  "pet",  "pr",  "rmax",  "rmin",  "sph",  "srad",  "th",  "tmmn",  "tmmx",  "vpd",  "vs")}}

\item{start_year}{An integer of year to start download data for (minimum possible is 1979).}

\item{end_year}{An integer of year to end data download (max possible is the current year).}
}
\value{
NA, data are saved to \code{data_dir/raw}
}
\description{
Create a data directory for creating climate normals and download
gridMET data to it for a specified period of record. This function requires
\code{wget} to be globally available on your system.
}
\examples{
\dontrun{
fetch_gridmet("~/data_dir", c("tmmx", "tmmn", "pr"), 1991, 2020)
}
}
