% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nclimgrid.R
\name{fetch_nclimgrid}
\alias{fetch_nclimgrid}
\title{fetch_nclimgrid}
\usage{
fetch_nclimgrid(
  data_dir,
  start_year = 1950,
  end_year = 2015,
  variables = c("tmin", "tmax", "tavg", "prcp"),
  cores = NULL
)
}
\arguments{
\item{data_dir}{Path to the directory where data will be saved.}

\item{start_year}{The year to start downloading data from. Defaults to 1951,
the earliest year daily data are available for.}

\item{end_year}{The year to end data downloads for. Defaults to 2015.}

\item{variables}{A vector of variables to download. Options are
c("tmin", "tmax", "tavg", "prcp")}

\item{cores}{Number of cores to use if you want to parallelize downloads.
If left as null, a single core will be used.}
}
\value{
A \code{tibble::tibble()} with paths to each of the new nclimgrid rasters.
}
\description{
Download daily nClimGrid data from NOAA.
}
\examples{
\dontrun{
fetch_nclimgrid("~/data/nclimgrid", 2010, 2011, variables = c("tmin", "tmax"))
}
}
