% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_to_reference}
\alias{filter_to_reference}
\title{filter_to_reference}
\usage{
filter_to_reference(r, start_year, end_year)
}
\arguments{
\item{r}{The \code{terra::rast} object to filter.}

\item{start_year}{An integer representing the year to start the filter.}

\item{end_year}{An integer representing the year to end the filter}
}
\value{
A \code{terra::rast} with the years filtered to match the reference period.
}
\description{
Filter a timeseries \code{terra::rast} so that all layers are within
a reference period.
}
\examples{
\dontrun{
r <- "https://data.climate.umt.edu/mca/cmip/ACCESS-ESM1-5_historical_r1i1p1f1_pr.tif"
r <- terra::rast(r)

filter_to_reference(r, 1981, 2010)
}
}
