% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normals.R
\name{gamma_from_rast}
\alias{gamma_from_rast}
\title{Calculate gamma normals for an input raster timeseries.}
\usage{
gamma_from_rast(x, out_dir = NULL, descriptor = "annual", rast_out = FALSE)
}
\arguments{
\item{x}{A \code{terra::rast} monthly or annual timeseries that will be used to
calculate gamma normals.}

\item{out_dir}{Optional. The directory to save the results out to.}

\item{descriptor}{The prefix to use on the filename as an identifier.}
}
\value{
A named list of \code{terra::rast} summarized to the following items:
'mean', 'median', mode', 'variance', 'alpha', 'beta'.
}
\description{
Calculate gamma normals for an input raster timeseries.
}
\examples{
a = 1
}
