% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_tables.R
\name{group_seasonally}
\alias{group_seasonally}
\title{group_seasonally}
\usage{
group_seasonally(r, start_year, end_year)
}
\arguments{
\item{r}{The \code{terra::rast} timeseries to create the table for.}

\item{start_year}{The start year to filter the timeseries by. Any years prior
to this year will be removed}

\item{end_year}{The end year to filter the timeseries by. Any years after this
year will be removed.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} giving the seasons of each layer in a raster
}
\description{
Create a table giving seasons for each raster layer in a \code{terra::rast} timeseries.
}
\examples{
\dontrun{
# fill with example
1+1
}
}
