% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{make_boxplot_plot}
\alias{make_boxplot_plot}
\title{make_boxplot_plot}
\usage{
make_boxplot_plot(dat, ylab, title)
}
\arguments{
\item{dat}{A \code{tibble} returned by [\code{\link[=make_boxplot_data]{make_boxplot_data()}}].}

\item{ylab}{The label to apply to the y axis of the plot.}

\item{title}{The title to give the figure.}
}
\value{
A \code{ggplot2} object.
}
\description{
make_boxplot_plot
}
\examples{
\dontrun{
1+1
}
}
