% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_gridmet_tavg}
\alias{make_gridmet_tavg}
\title{make_gridmet_tavg}
\usage{
make_gridmet_tavg(tmmn_dir, tmmx_dir, tavg_dir)
}
\arguments{
\item{tmmn_dir}{The directory of data containing the raw gridMET tmmn .nc files.}

\item{tmmx_dir}{The directory of data containing the raw gridMET tmmx .nc files.}

\item{tavg_dir}{The directory to save the new gridMET tavg .nc files to.}
}
\description{
This function averages gridMET minimum and maximum temperature
products to create a gridMET average temperature. This function uses the raw
output from \code{fetch_gridmet} to create the average temperature product.
}
\examples{
\dontrun{
make_gridmet_tavg(
  "~/gridmet_data/raw/tmmn",
  "~/gridmet_data/raw/tmmx",
  "~/gridmet_data/raw/tavg"
)
}
}
