% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{make_heatmap_data}
\alias{make_heatmap_data}
\title{make_heatmap_data}
\usage{
make_heatmap_data(files, pattern, fun, shp, attr_id, tsfm = NULL)
}
\arguments{
\item{files}{A list of CMIP6 climate projection files.}

\item{pattern}{A string pattern to use to subset the above CMIP6 files.}

\item{fun}{A function used to aggregate monthly CMIP6 data to a seasonal or annual timescale.}

\item{shp}{An \code{sf} object that will be used to summarise the CMIP6 data.}

\item{attr_id}{The column in \code{shp} that will be used to aggregate by}

\item{tsfm}{A function to apply to the data (e.g., a function that converts from
degrees C to degrees F).}
}
\value{
An \code{tibble} object that will be used to plot a heatmap of projected mid and
end of century changes in a given climate variable
}
\description{
make_heatmap_data
}
\examples{
\dontrun{
1+1
}
}
