% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{make_heatmap_plot}
\alias{make_heatmap_plot}
\title{make_heatmap_plot}
\usage{
make_heatmap_plot(dat, hot = TRUE, title_txt)
}
\arguments{
\item{dat}{A \code{tibble} returned by [\code{\link[=make_heatmap_data]{make_heatmap_data()}}].}

\item{hot}{A boolean specifying whether to use a hot or cool colorscale.
Defaults to TRUE.}

\item{title_txt}{The title to give to the plot.}
}
\value{
A \code{ggplot2} object.
}
\description{
make_heatmap_plot
}
\examples{
\dontrun{
1+1
}
}
