% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{make_map_data}
\alias{make_map_data}
\title{make_map_data}
\usage{
make_map_data(
  files,
  pattern,
  fun,
  shp,
  attr_id,
  tsfm = NULL,
  proj = normals::mt_state_plane
)
}
\arguments{
\item{files}{A list of CMIP6 climate projection files.}

\item{pattern}{A string pattern to use to subset the above CMIP6 files.}

\item{fun}{A function used to aggregate monthly CMIP6 data to a seasonal or annual timescale.}

\item{shp}{An \code{sf} object that will be used to summarise the CMIP6 data.}

\item{attr_id}{The column in \code{shp} that will be used to aggregate by}

\item{tsfm}{A function to apply to the data (e.g., a function that converts from
degrees C to degrees F).}

\item{proj}{A WKT string or [\code{\link[sf:st_crs]{sf::st_crs()}}] object giving a projection to use
for the output data.}
}
\value{
An \code{sf} object that will be used to plot a map of projected mid and
end of century changes in a given climate variable
}
\description{
make_map_data
}
\examples{
\dontrun{
1+1
}
}
