% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_tables.R
\name{make_pr_seasonal_raster}
\alias{make_pr_seasonal_raster}
\title{make_pr_seasonal_raster}
\usage{
make_pr_seasonal_raster(
  pr,
  out_dir,
  start_year,
  end_year,
  dataset = "nclimgrid"
)
}
\arguments{
\item{pr}{A \code{terra::rast} timeseries of monthly precipitation totals.}

\item{out_dir}{The directory to save the seasonal rasters out to.}

\item{start_year}{The start year of the reference period to filter input data by.}

\item{end_year}{The end year of the reference period to filter input data by.}
}
\value{
Results are saved to an output directory.
}
\description{
make_pr_seasonal_raster
}
\examples{
\dontrun{
# fill with example
1+1
}
}
