% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_trends.R
\name{make_seasonal_trend_table}
\alias{make_seasonal_trend_table}
\title{make_seasonal_trend_table}
\usage{
make_seasonal_trend_table(
  r,
  shp,
  attr_id,
  fun,
  start_year,
  end_year,
  annual = FALSE
)
}
\arguments{
\item{r}{A timeseries \code{terra::rast} of an annual climate variable.}

\item{shp}{An \code{sf} object with regions to summarise \code{r} by.}

\item{attr_id}{The column in \code{shp} to use as a name.}

\item{fun}{The function to summarise each region in \code{shp} by. Defaults to "mean".}

\item{start_year}{The year to filter the start or the data by.}

\item{end_year}{The year to filter the end of the data by.}

\item{annual}{A boolean specifying whether data should be aggregated annually.
Defaults to False, which aggregates data seasonally.}
}
\value{
A table giving the decadal rate of change of the climate variable.
}
\description{
make_seasonal_trend_table
}
\examples{
\dontrun{
1+1
}
}
