% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_tables.R
\name{make_temp_seasonal_rasters}
\alias{make_temp_seasonal_rasters}
\title{make_temp_seasonal_rasters}
\usage{
make_temp_seasonal_rasters(tmmn, tmmx, out_dir, start_year, end_year)
}
\arguments{
\item{tmmn}{A \code{terra::rast} timeseries of minimum temperature.}

\item{tmmx}{A \code{terra::rast} timeseries of maximum temperature.}

\item{out_dir}{The directory to save the seasonal rasters out to.}

\item{start_year}{The start year of the reference period to filter input data by.}

\item{end_year}{The end year of the reference period to filter input data by.}
}
\value{
Nothing, results are saved to an output directory.
}
\description{
make_temp_seasonal_rasters
}
\examples{
\dontrun{
# fill with example
1+1
}
}
