% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_trends.R
\name{make_trend_fig}
\alias{make_trend_fig}
\title{make_trend_fig}
\usage{
make_trend_fig(
  r,
  shp,
  attr_id,
  fun = "mean",
  start_year = 1951,
  end_year = 2015,
  ylab = ""
)
}
\arguments{
\item{r}{A timeseries \code{terra::rast} of an annual climate variable.}

\item{shp}{An \code{sf} object with regions to summarise \code{r} by.}

\item{attr_id}{The column in \code{shp} to use as a name.}

\item{fun}{The function to summarise each region in \code{shp} by. Defaults to "mean".}

\item{start_year}{The year to filter the start or the data by.}

\item{end_year}{The year to filter the end of the data by.}

\item{ylab}{The label to use for the y-axis of the figure.}
}
\value{
A \code{ggplot2} figure of change in temperature per decade per region in shp.
}
\description{
make_trend_fig
}
\examples{
\dontrun{
1+1
}
}
