% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mt}
\alias{mt}
\title{\code{sf} object of the outline of Montana.}
\format{
A geometry set with one feature.
}
\source{
\url{http://ftp.geoinfo.msl.mt.gov/Data/Spatial/MSDI/AdministrativeBoundaries/MontanaCounties.zip}
}
\usage{
mt
}
\description{
A dataset outlining Montana to use as an example in the repo.
The shapefile was taken from the \code{mcor} R package.
}
\keyword{datasets}
