% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mt_state_plane}
\alias{mt_state_plane}
\title{The Montana State Plane Coordinate Reference System
This object was duplicated from the \href{https://github.com/mt-climate-office/mcor}{mcor}
package.}
\format{
An object of class \link[sf:st_crs]{sf::crs} of length 2.
\itemize{
\item \strong{epsg} — an integer (102300) representing the (\href{http://epsg.org/}{EPSG}) code
\item \strong{proj4string} — a character string representing the associated \href{http://proj4.org/}{proj.4} code
}
}
\source{
\url{https://epsg.io/102300}
}
\usage{
mt_state_plane
}
\description{
An object of class \link[sf:st_crs]{sf::crs} defining the
\emph{NAD 1983 HARN StatePlane Montana FIPS 2500} projection
coordinate system (\href{https://epsg.io/102300}{EPSG: 102300}).
}
\keyword{datasets}
