% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spat_summary}
\alias{spat_summary}
\title{spat_summary}
\usage{
spat_summary(rasts, shp, attr_id = NULL, name_to = "timescale", fun, ...)
}
\arguments{
\item{rasts}{A \code{terra::rast} that zonal statistics will be derived from.}

\item{shp}{An \code{sf} object that will be used to derive zonal statistics.}

\item{attr_id}{A string of the column name in \code{shp} describing each subregion.}

\item{name_to}{The desired name of the descriptor column in the output dataframe.}

\item{fun}{The function to apply in the zonal statistics.}

\item{...}{Any additional arguments that should be passed to fun}
}
\value{
A dataframe of the aggregated statistics for each region in \code{shp}.
}
\description{
spat_summary
}
\examples{
\dontrun{
mt <- normals::mt \%>\% dplyr::mutate(name = "MT")
r <- normals::read_from_server(
"https://data.climate.umt.edu/mca/cmip/ACCESS-ESM1-5_historical_r1i1p1f1_above90.tif"
)

spat_summary(r, shp, "name", "time", "mean")
}
}
