% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_shp}
\alias{to_shp}
\title{to_shp}
\usage{
to_shp(x, shp, proj)
}
\arguments{
\item{x}{A \code{terra::sf} that will be converted into polygons.}

\item{shp}{The boundaries to intersect the polygonized raster with.}

\item{proj}{The projection to transform the \code{terra::rast} into. Can either be
an \code{sf::st_crs} object or a string of the WKT.}
}
\value{
A polygonized version of the \code{terra::rast} input.
}
\description{
Convert a \code{terra::rast} object to \code{sf} polygons. This allows you
to use \link[ggplot2:ggsf]{ggplot2::geom_sf} to nicely plot the data.
}
\examples{
\dontrun{
r <- "https://data.climate.umt.edu/mca/cmip/ACCESS-ESM1-5_historical_r1i1p1f1_pr.tif"
r <- terra::rast(r)

to_shp(r, shp=normals::mt, proj=normals::mt_state_plane)
}
}
