% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{make_boxplot_data}
\alias{make_boxplot_data}
\title{make_boxplot_data}
\usage{
make_boxplot_data(f, shp, attr_id, fun = "mean")
}
\arguments{
\item{f}{Preprocessed cmip6 data returned by [\link{cmip_monthly_to_change}].}

\item{shp}{An \code{sf} object that will be used to summarize the CMIP6 data.}

\item{attr_id}{The column in \code{shp} that will be used to aggregate by}
}
\value{
A \code{tibble} of annual changes in a variable according to mid and end of
century projections, aggregated by \code{attr_id} columns.
}
\description{
make_boxplot_data
}
\examples{
\dontrun{
1+1
}
}
