% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{make_heatmap_data}
\alias{make_heatmap_data}
\title{make_heatmap_data}
\usage{
make_heatmap_data(f, shp, attr_id = NULL)
}
\arguments{
\item{f}{An \code{.rds} file created by [\link{cmip_monthly_to_change}].}

\item{shp}{An \code{sf} object that will be used to summarise the CMIP6 data.}

\item{attr_id}{The column in \code{shp} that will be used to aggregate by, defaults to NULL.
If left blank, results will be aggregated across the entire domain.}
}
\value{
An \code{tibble} object that will be used to plot a heatmap of projected mid and
end of century changes in a given climate variable
}
\description{
make_heatmap_data
}
\examples{
\dontrun{
1+1
}
}
