% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cmip.R
\name{make_map_data}
\alias{make_map_data}
\title{make_map_data}
\usage{
make_map_data(f, shp, attr_id, proj = normals::mt_state_plane, fun = "mean")
}
\arguments{
\item{f}{The preprocessed .Rds file to read from.}

\item{shp}{The [\link{sf}] object to use to aggregate data by.}

\item{attr_id}{The column in \code{shp} that will be used to aggregate by}

\item{proj}{A WKT string or [\code{\link[sf:st_crs]{sf::st_crs()}}] object giving a projection to use
for the output data.}

\item{fun}{A function to aggreagte each region in \code{shp} by.}
}
\value{
An \code{sf} object that will be used to plot a map of projected mid and
end of century changes in a given climate variable
}
\description{
make_map_data
}
\examples{
\dontrun{
1+1
}
}
