% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_as_cog}
\alias{write_as_cog}
\title{write_as_cog}
\usage{
write_as_cog(x, filename)
}
\arguments{
\item{filename}{}
}
\description{
A wrapper around \code{terra::writeRaster} to write a \code{terra::rast}
to disk as a \href{https://www.cogeo.org/}{Cloud Optimized GeoTIFF}.
}
